+++ level/integration/src/main/java/com/fitbur/testify/integration/injector/IntegrationRealServiceInjector.java
@@ -15,6 +15,7 @@
  */
 package com.fitbur.testify.integration.injector;
 
+import com.fitbur.testify.Real;
 import com.fitbur.testify.TestContext;
 import com.fitbur.testify.TestInjector;
 import com.fitbur.testify.TestReifier;
@@ -26,6 +27,7 @@ import java.lang.reflect.Parameter;
 import java.lang.reflect.Type;
 import java.util.Collection;
 import java.util.Map;
+import javax.inject.Inject;
 
 /**
  * An integration test injector implementation that injects fields annotated
@@ -39,32 +41,33 @@ public class IntegrationRealServiceInjector implements TestInjector {
 
     private final TestContext context;
     private final TestReifier testReifier;
-    private final FieldDescriptor fieldDescriptor;
     private final Object[] arguments;
 
     public IntegrationRealServiceInjector(TestContext context,
             TestReifier testReifier,
-            FieldDescriptor fieldDescriptor,
             Object[] arguments) {
         this.context = context;
         this.testReifier = testReifier;
-        this.fieldDescriptor = fieldDescriptor;
         this.arguments = arguments;
     }
 
     @Override
-    public void inject() {
+    public void inject(FieldDescriptor descriptor) {
+        if (!descriptor.hasAnyAnnotation(Real.class, Inject.class)) {
+            return;
+        }
+
         Map<DescriptorKey, ParameterDescriptor> parameterDescriptors = context.getParamaterDescriptors();
-        Type fieldType = fieldDescriptor.getGenericType();
-        String fieldName = fieldDescriptor.getName();
+        Type fieldType = descriptor.getGenericType();
+        String fieldName = descriptor.getName();
         DescriptorKey descriptorKey = new DescriptorKey(fieldType, fieldName);
 
         //if there is a parameter descriptor that matches the field then lets use that
         if (parameterDescriptors.containsKey(descriptorKey)) {
-            ParameterDescriptor descriptor = parameterDescriptors.get(descriptorKey);
-            Integer index = descriptor.getIndex();
+            ParameterDescriptor paramDescriptor = parameterDescriptors.get(descriptorKey);
+            Integer index = paramDescriptor.getIndex();
 
-            Object instance = testReifier.reifyField(fieldDescriptor, descriptor);
+            Object instance = testReifier.reifyField(descriptor, paramDescriptor);
             arguments[index] = instance;
         } else {
             TypeToken token = TypeToken.of(fieldType);
@@ -80,7 +83,7 @@ public class IntegrationRealServiceInjector implements TestInjector {
                 }
 
                 if (token.isSubtypeOf(paramType)) {
-                    Object instance = testReifier.reifyField(fieldDescriptor, paramDescriptor);
+                    Object instance = testReifier.reifyField(descriptor, paramDescriptor);
                     arguments[index] = instance;
                     break;
                 }